refactor node into logic and graphics
