+++ node_editor/gui/common.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class Node_Status(Enum):
+    CLEAN = 1
+    DIRTY = 2
+    ERROR = 3
